+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Stability fixes
 - change type of `transaction_id` to string
 - `auth` parameter of subscription changed to `accessKey`
 - invalid `accessKey` treated by subscribe as a valid key
+- all internal errors are logged as is but converted to `Service temporary unavailable` before sending to client
+- server side stack traces are truncated before sending to client 
 
 ## 0.24.3 – Mar 2, 2020
 ### Featured