+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Release Notes
 All notable changes to this project will be documented in this file.
 
+## 0.24.4 – Mar 12, 2020
+### Featured
+Security fixes
+
+### New
+- `--mam-access-keys` and `MAM_ACCESS_KEYS` config to protect mam endpoint
+- all queries and mutations inside single GraphQL request must use the same access key
+
+### Fixed
+- change type of `transaction_id` to string
+- `auth` parameter of subscription changed to `accessKey`
+- invalid `accessKey` treated by subscribe as a valid key
+- all internal errors are logged as is but converted to `Service temporary unavailable` before sending to client
+- server side stack traces are truncated before sending to client 
+
 ## 0.24.3 – Mar 2, 2020
 ### Featured
 Stability fixes